feat(route/threads): ditch puppeteer
