+++ website/docs/routes/social-media.mdx
@@ -889,7 +889,7 @@ Due to Telegram restrictions, some channels involving pornography, copyright, an
 
 ### User timeline {#threads-user-timeline}
 
-<Route author="ninboy" path="/threads/:user/:routeParams?" example="/threads/zuck" radar="1" rssbud="1" puppeteer="1" paramsDesc={['', '']}>
+<Route author="ninboy" path="/threads/:user/:routeParams?" example="/threads/zuck" radar="1" rssbud="1" paramsDesc={['Username', 'Extra parameters, see the table below']}>
 
 Specify options (in the format of query string) in parameter `routeParams` to control some extra features for threads
